+++ src/main/kotlin/org/xbery/artbeams/web/FreeProductController.kt
@@ -335,16 +335,14 @@ class FreeProductController(
         try {
             val productAuthor = userService.findById(product.createdBy)
             if (productAuthor != null) {
-                if (productAuthor.email.isNotEmpty()) {
-                    val subject =
-                        normalizationHelper.removeDiacriticalMarks("User ${user.firstName} ${user.lastName} downloaded ${product.title}")
-                    val body =
-                        normalizationHelper.removeDiacriticalMarks(
-                            "User ${user.firstName} ${user.lastName}/${user.login} " +
-                                    "has downloaded product ${product.title}."
-                        )
-                    mailSender.sendMailWithText(productAuthor.email, subject, body)
-                }
+                val subject =
+                    normalizationHelper.removeDiacriticalMarks("User ${user.firstName} ${user.lastName} downloaded ${product.title}")
+                val body =
+                    normalizationHelper.removeDiacriticalMarks(
+                        "User ${user.firstName} ${user.lastName}/${user.login} " +
+                                "has downloaded product ${product.title}."
+                    )
+                mailSender.sendMailWithText(productAuthor.login, subject, body)
             }
         } catch (ex: Exception) {
             logger.error("Error while sending product downloaded notification: ${ex.message}", ex)